+++ backend/local_factory_test.go
@@ -23,12 +23,16 @@ func TestLocalFactory(t *testing.T) {
 
 	testGenericFactory(t, f)
 
-	_, err := f.New("", map[string]interface{}{})
+	_, err := f.New(map[string]interface{}{})
 	if err == nil {
 		t.Error("expected error")
 	}
 
-	_, err = f.New("toto", map[string]interface{}{})
+	_, err = f.New(map[string]interface{}{
+		"local": map[string]interface{}{
+			"path": "/tmp/store.scrt",
+		},
+	})
 	if err != nil {
 		t.Error(err)
 	}