+++ cmd/root_test.go
@@ -47,12 +47,6 @@ func TestRootCmd(t *testing.T) {
 		t.Fatal("expected error")
 	}
 
-	viper.Set(configKeyLocation, "")
-	err = RootCmd.PersistentPreRunE(RootCmd, []string{})
-	if err == nil {
-		t.Fatal("expected error")
-	}
-
 	viper.Set(configKeyPassword, "")
 	err = RootCmd.PersistentPreRunE(RootCmd, []string{})
 	if err != nil {