+++ cmd/unset_test.go
@@ -39,7 +39,6 @@ func TestUnsetCmd(t *testing.T) {
 	viper.Reset()
 	viper.Set(configKeyPassword, password)
 	viper.Set(configKeyStorage, "mock")
-	viper.Set(configKeyLocation, "location")
 
 	s := store.NewStore()
 	err := s.Set("hello", []byte("world"))
@@ -78,7 +77,6 @@ func TestUnsetCmdNoValue(t *testing.T) {
 	viper.Reset()
 	viper.Set(configKeyPassword, password)
 	viper.Set(configKeyStorage, "mock")
-	viper.Set(configKeyLocation, "location")
 
 	s := store.NewStore()
 	data, err := store.WriteStore([]byte(password), s)
@@ -111,7 +109,6 @@ func TestUnsetCmdNotExists(t *testing.T) {
 	viper.Reset()
 	viper.Set(configKeyPassword, "toto")
 	viper.Set(configKeyStorage, "mock")
-	viper.Set(configKeyLocation, "location")
 
 	mockBackend.EXPECT().Exists().Return(false, nil)
 
@@ -136,7 +133,6 @@ func TestUnsetCmdFailedLoad(t *testing.T) {
 	viper.Reset()
 	viper.Set(configKeyPassword, "toto")
 	viper.Set(configKeyStorage, "mock")
-	viper.Set(configKeyLocation, "location")
 
 	mockBackend.EXPECT().Exists().Return(true, nil)
 	mockBackend.EXPECT().Load().Return(nil, fmt.Errorf("error"))
@@ -164,7 +160,6 @@ func TestUnsetCmdFailedSave(t *testing.T) {
 	viper.Reset()
 	viper.Set(configKeyPassword, password)
 	viper.Set(configKeyStorage, "mock")
-	viper.Set(configKeyLocation, "location")
 
 	s := store.NewStore()
 	data, err := store.WriteStore([]byte(password), s)